+++ test/test_scrabble_game.py
@@ -31,7 +31,7 @@ class TestScrabbleGame(unittest.TestCase):
         scrabble_game.next_turn()
         assert scrabble_game.current_player == scrabble_game.players[0]
     
-    def test_word_validation_only_tiles_user_and_board(self):
+    """def test_word_validation_only_tiles_user_and_board(self):
         scrabble_game = ScrabbleGame(players_count=3)
         scrabble_game.current_player = scrabble_game.players[2]
         scrabble_game.current_player.bag_tiles = (Tiles("A", 1), Tiles("H", 2), Tiles("L", 3), Tiles("O", 4))
@@ -93,7 +93,7 @@ class TestScrabbleGame(unittest.TestCase):
         orientation = "H"
         validation = scrabble_game.validate_word(word, location, orientation)
         self.assertEqual(validation, True)
-    
+    """
     def test_put_simple_word_h(self):
         scrabble_game = ScrabbleGame(players_count=3)
         scrabble_game.current_player = scrabble_game.players[2]
@@ -164,6 +164,16 @@ class TestCalculateWordValue(unittest.TestCase):
         location = (0,0)
         value = scrabble_game.calculate_word_value(word, location, orientation)
         self.assertEqual(value, 18)
+
+    def test_with_letter_word_multiplier_2_word(self):
+        scrabble_game = ScrabbleGame(players_count=3)
+        scrabble_game.current_player = scrabble_game.players[2]
+        scrabble_game.current_player.tiles = [Tiles("M", 2), Tiles("A", 1), Tiles("S", 2), Tiles("A", 1), Tiles("Y", 4), Tiles("O", 1), Tiles("N", 1), Tiles("E", 1)]
+        word = "MAYONESA"
+        orientation = "V"
+        location = (0,0)
+        value = scrabble_game.calculate_word_value(word, location, orientation)
+        self.assertEqual(value, 126)
 
     def test_with_letter_word_multiplier_no_active(self):
         scrabble_game = ScrabbleGame(players_count=3)